projects/discordbot/pyproject.toml
[tool.poetry]
name = "discordbot"
version = "0.1.0"
description = "Discord question answer bot"
authors = ["Vaibhav Bhandari <vaibhavb@gmail.com>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^2.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
